Optimising code laodout + adding more cecurity and bether tekst

+++ classes/projects.php
@@ -1,13 +1,20 @@
 <?php
 
-class Projects {
+/**
+ * projects
+ */
+class Projects
+{
 
     /**
      * @function
      * load projects on limit
+     * @param int $limit
+     * @return array|false
      */
 
-    public static function loadprojects($limit = 3) {
+    public static function loadprojects($limit = 3)
+    {
         $results = database::getRows('projecten', ['removed'], 's', [0], 'date DESC LIMIT ' . $limit);
         if ($results) {
             return $results;
@@ -19,9 +26,12 @@ class Projects {
     /**
      * @function
      * load project by id
+     * @param int $id
+     * @return array|false|mixed
      */
 
-    public static function loadproject($id = 0) {
+    public static function loadproject($id = 0)
+    {
         if ($id == 0) {
             return false;
         }
@@ -36,34 +46,57 @@ class Projects {
     /**
      * @function
      * start and end of the add project function
+     * @param $file
+     * @param $title
+     * @param $pgit
+     * @param $plink
+     * @return string|void
      */
 
-    public static function addproject($target_dir, $target_file, $imageFileType, $filename, $name, $continue, $files, $title, $pgit, $plink) {
-        $img = Projects::uploadimg($target_dir, $target_file, $imageFileType, $files);
-        if ($filename and $name) {
-            $zip = Projects::uploadzip($filename, $name, $continue, $files);
+    public static function addproject($file, $title, $pgit, $plink)
+    {
+        $continue = false;
+        if ($plink !== "") {
+            $continue = true;
+            $filename = false;
+            $name = false;
+        } else {
+            $filename = $file["zip_file"]["name"];
+            $name = explode(".", $filename);
+            $continue = strtolower($name[1]) == 'zip' ? true : false;
+        }
 
+        $target_dir = "img/";
+        $target_file = $target_dir . uniqid() . "_" . basename($file["img"]["name"]);
+        $imageFileType = strtolower(pathinfo($target_file, PATHINFO_EXTENSION));
+        if (
+            $imageFileType != "jpg" && $imageFileType != "png" && $imageFileType != "jpeg"
+            && $imageFileType != "gif"
+        ) {
+            return '<script>alert("Your uploaded file is not a img");</script>';
+        } elseif (!$continue) {
+            return '<script>alert("Your uploaded file is not a zip");</script>';
+        }
+
+        $img = Projects::uploadimg($target_dir, $target_file, $imageFileType, $file);
+        if ($filename and $name) {
+            $zip = Projects::uploadzip($filename, $name, $continue, $file);
         } else {
             $zip = true;
         }
 
         if (!$img) {
-            echo "<script>alert('Sorry, there was an error uploading your img');</script>";
+            return "<script>alert('Sorry, there was an error uploading your img');</script>";
         } elseif (!$zip) {
-            echo "<script>alert('Your .zip file was not uploaded and unpacked');</script>";
+            return "<script>alert('Your .zip file was not uploaded and unpacked');</script>";
         } else {
-            $git = "";
-            if ($pgit == "") {
-                $git = "empty";
-            } else {
-                $git = $_POST["git"];
-            }
+            $git = $pgit;
 
             $link = "";
             if ($plink == "") {
                 $link = substr($zip, 0, -4) . "/index";
             } else {
-                $link = $_POST["link"];
+                $link = $plink;
             }
             projects::dbaddproject($title, $git, $link, $img);
         }
@@ -72,29 +105,42 @@ class Projects {
     /**
      * @function
      * update the project
+     * @param $id
+     * @param $name
+     * @param $github
      */
-    public static function update($id = 0, $name, $github){
+    public static function update($id = 0, $name, $github)
+    {
         database::update('projecten', $id, ['name', 'github'], 'ss', [$name, $github]);
     }
 
     /**
      * @function
      * add project to databse
+     * @param $name
+     * @param $git
+     * @param $link
+     * @param $file
      */
-
-    public static function dbaddproject($name, $git, $link, $file) {
+    public static function dbaddproject($name, $git, $link, $file)
+    {
         Database::add('projecten', ['name', 'github', 'path', 'img', 'date'], 'sssss', [$name, $git, $link, $file, date('y-m-d h:m:s')]);
     }
 
     /**
      * @function
      * Upload the given img
+     * @param $target_dir
+     * @param $target_file
+     * @param $imageFileType
+     * @param $files
+     * @return array|false
      */
-
-    private static function uploadimg($target_dir, $target_file, $imageFileType, $files) {
+    private static function uploadimg($target_dir, $target_file, $imageFileType, $files)
+    {
         $uploadOk = 1;
 
-        // Check if image file is a actual image or fake image
+        // Check if image file is actual image or fake image
         if (isset($_POST["submit"])) {
             $check = getimagesize($files["img"]["tmp_name"]);
             if ($check !== false) {
@@ -106,7 +152,6 @@ class Projects {
 
         // Check if file already exists
         if (file_exists($target_file)) {
-            echo "<script>alert('file already exists');</>";
             $uploadOk = 0;
         }
 
@@ -123,12 +168,18 @@ class Projects {
     /**
      * @function
      * upload the given zip and unpack
+     * @param $filename
+     * @param $name
+     * @param $continue
+     * @param $files
+     * @return array|false
      */
 
-    private static function uploadzip($filename, $name, $continue, $files) {
+    private static function uploadzip($filename, $name, $continue, $files)
+    {
         if ($files["zip_file"]["name"]) {
             $source = $files["zip_file"]["tmp_name"];
-            $type   = $files["zip_file"]["type"];
+            $type = $files["zip_file"]["type"];
 
             $accepted_types = array('application/zip', 'application/x-zip-compressed', 'multipart/x-zip', 'application/x-compressed');
             foreach ($accepted_types as $mime_type) {
@@ -139,14 +190,14 @@ class Projects {
             }
 
             if (!$continue) {
-                echo "<script>alert('The file you are trying to upload is not a .zip file. Please try again');</script>";
+                return "<script>alert('The file you are trying to upload is not a .zip file. Please try again');</script>";
             } else {
                 $target_path = 'project/' . $filename;
                 if (move_uploaded_file($source, $target_path)) {
                     $zip = new ZipArchive();
-                    $x   = $zip->open($target_path);
+                    $x = $zip->open($target_path);
                     if ($x === true) {
-                        $zip->extractTo('project/');
+                        $zip->extractTo('../project/');
                         $zip->close();
                         unlink($target_path);
                     }
@@ -161,33 +212,53 @@ class Projects {
     /**
      * @function
      * Softdelete the project
+     * @param $id
      */
-    public static function sdeleteproject($id){
+    public static function sdeleteproject($id)
+    {
         database::update('projecten', $id, ['removed'], 's', [1]);
     }
-    
-    public static function harddelete($id){
-        $account = Projects::loadproject($id);
-        $path = (substr($account['path'], 0, -6));
-        unlink($account['img']);
 
-        function deleteDirectory($path)
-        {
-            if (is_dir($path)) {
-                $objects = scandir($path);
-                foreach ($objects as $object) {
-                    if ($object != "." && $object != "..") {
-                        if (filetype($path . DIRECTORY_SEPARATOR . $object) == "dir") {
-                            deleteDirectory($path . DIRECTORY_SEPARATOR . $object);
-                        } else {
-                            unlink($path . DIRECTORY_SEPARATOR . $object);
+    /**
+     * @function
+     * full delete project
+     */
+    public static function delete()
+    {
+        $results = database::getRows('projecten', ['removed'], 's', [1]);
+        foreach ($results as $result) {
+            $date = date_create($result['updated']);
+            date_add($date, date_interval_create_from_date_string("6 day"));
+            $max_date = date_format($date, "y-m-d");
+            if ($max_date <= date("y-m-d")) {
+                Database::delete('projecten', $result['id']);
+
+                $path = (substr($result['path'], 0, -6));
+                unlink($result['img']);
+
+                function deleteDirectory($path)
+                {
+                    if (is_dir($path)) {
+                        $objects = scandir($path);
+                        foreach ($objects as $object) {
+                            if ($object != "." && $object != "..") {
+                                if (filetype($path . DIRECTORY_SEPARATOR . $object) == "dir") {
+                                    deleteDirectory($path . DIRECTORY_SEPARATOR . $object);
+                                } else {
+                                    unlink($path . DIRECTORY_SEPARATOR . $object);
+                                }
+                            }
                         }
+                        rmdir($path);
                     }
                 }
-                reset($objects);
-                rmdir($path);
+
+                deleteDirectory($path);
             }
+
         }
-        deleteDirectory($path);
+
     }
 }
+
+
